// part9/Patientor/Patientor/src/types.ts
export interface Diagnosis {
  code: string
  name: string
  latin?: string
}

export enum Gender {
  Male = 'male',
  Female = 'female',
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
// export interface Entry {}

export interface Patient {
  id: string
  name: string
  dateOfBirth: string
  ssn: string
  gender: Gender
  occupation: string
  entries: Entry[]
}

export type PatientOutput = Omit<Patient, 'ssn' | 'entries'>;

export type NewPatient = Omit<Patient, 'id'>;

interface BaseEntry {
  id: string
  type: EntryType
  description: string
  date: string
  specialist: string
  diagnosisCodes?: Array<Diagnosis['code']>
}

export enum HealthCheckRating {
  'Healthy' = 0,
  'LowRisk' = 1,
  'HighRisk' = 2,
  'CriticalRisk' = 3,
}

interface HealthCheckEntry extends BaseEntry {
  type: EntryType.HealthCheck
  healthCheckRating: HealthCheckRating
}

export enum EntryType {
  Hospital = 'Hospital',
  OccupationalHealthcare = 'OccupationalHealthCare',
  HealthCheck = 'HealthCheck',
}

export interface Discharge {
    date: string
    criteria: string
}

interface HospitalEntry extends BaseEntry {
  type: EntryType.Hospital
  discharge: Discharge
}

export interface SickLeave {
  startDate: string
  endDate: string
}

interface OccupationalHealthcareEntry extends BaseEntry {
  type: EntryType.OccupationalHealthcare
  employerName: string
  sickLeave?: SickLeave
}

export type Entry = HospitalEntry | OccupationalHealthcareEntry | HealthCheckEntry;

// eslint-disable-next-line @typescript-eslint/no-explicit-any
type UnionOmit<T, K extends keyof any> = T extends any ? Omit<T, K> : never;

export type baseEntry = Omit<BaseEntry, 'id'>;

export type NewEntry = UnionOmit<Entry, 'id'>;